Fix the socat command.
pbench-fio assumes that if it can ping the remote that it is up and
ready to accept connections.  But the tunneling mechanism that we use
with socat always accepts connections locally even if the remote is
not ready to accept connections.  socat has no way to refuse
connections if the remote rejects a connection attempt.  The
workaround is to add the ,forever option to the remote so that it will
retry until connection succeeds.

This will not do the right thing if the remote entirely fails to come
up, but it's the best workaround available at this point.