Enable use _idname() as a deterministic random seed by using child index rather than process ID.
This will not work correctly for deployments or replicasets, where the
podnames have random hashes appended, so those deployment types should
not be used where deterministic IDs are required.

Also make child_idx private; the index is passed to each subprocess so
it is not necessary that clients have direct access to this member.