Fix the timestamp calculation at worker startup.
The startup attempts to determine the skew between the sync pod time and the worker pod time by determining the round trip time as seen by the worker (time to send request and receive reply) and the round trip time as seen remotely (time between receiving the request and servicing it).  The worker-side round trip time was calculated by taking the time stamp before and after the request to the sync pod.  That was being done incorrectly, because if the sync pod was not servicing requests when the initial request was made from the client, the time before the sync pod became responsive was considered part of the round trip time.  The fix is to not start that timer until a successful connection to the sync pod is made.